refactor: decompose workflow macro modules (stage 1) (#376)
