+++ lib/workflow-macros.ts
@@ -13,6 +13,19 @@
 
 import crypto from 'node:crypto';
 import { EventEmitter } from 'node:events';
+import { PREBUILT_WORKFLOWS, registerDefaultCommandsAndShortcuts } from './workflow-macros-defaults';
+import {
+  cloneValue,
+  estimateTotalUnits,
+  evaluateCondition,
+  getPathValue,
+  resolveBatchItems,
+  resolveTemplateValue,
+} from './workflow-macros-helpers';
+import { WorkflowRecorder } from './workflow-recorder';
+
+export { PREBUILT_WORKFLOWS };
+export { WorkflowRecorder };
 
 // ============================================================================
 // Shared Types
@@ -854,112 +867,6 @@ export class BatchWorkflowProcessor {
 }
 
 // ============================================================================
-// Workflow Recorder
-// ============================================================================
-
-export class WorkflowRecorder {
-  private draft?: WorkflowDefinition;
-  private readonly stepOrder: string[] = [];
-
-  start(config: { name: string; description: string; version?: number; tags?: string[] }): void {
-    this.draft = {
-      version: config.version ?? 1,
-      name: config.name,
-      description: config.description,
-      tags: config.tags,
-      steps: [],
-    };
-    this.stepOrder.length = 0;
-  }
-
-  recordStep(step: WorkflowStepDefinition): void {
-    this.assertStarted();
-
-    const clone = cloneValue(step) as WorkflowStepDefinition;
-    this.draft!.steps.push(clone);
-    this.stepOrder.push(clone.id);
-  }
-
-  recordCommand(stepId: string, command: WorkflowCommandReference, options?: {
-    description?: string;
-    mode?: WorkflowStepMode;
-    retry?: RetryPolicy;
-    onError?: WorkflowOnError;
-  }): void {
-    this.assertStarted();
-
-    const existing = this.draft!.steps.find((s) => s.id === stepId && s.type !== 'batch') as WorkflowCommandStepDefinition | undefined;
-    if (existing) {
-      existing.commands.push(cloneValue(command) as WorkflowCommandReference);
-      return;
-    }
-
-    const step: WorkflowCommandStepDefinition = {
-      id: stepId,
-      description: options?.description,
-      type: 'commands',
-      mode: options?.mode,
-      retry: options?.retry,
-      onError: options?.onError,
-      commands: [cloneValue(command) as WorkflowCommandReference],
-    };
-
-    this.draft!.steps.push(step);
-    this.stepOrder.push(stepId);
-  }
-
-  stop(): WorkflowDefinition {
-    this.assertStarted();
-
-    const out = cloneValue(this.draft) as WorkflowDefinition;
-    this.draft = undefined;
-    this.stepOrder.length = 0;
-    return out;
-  }
-
-  static fromRun(
-    result: WorkflowRunResult,
-    config: { name?: string; description?: string; version?: number; tags?: string[] } = {}
-  ): WorkflowDefinition {
-    const steps: WorkflowStepDefinition[] = [];
-
-    for (const step of result.steps) {
-      if (!step.commands.length) continue;
-
-      const commands: WorkflowCommandReference[] = step.commands.map((cmd) => ({
-        command: cmd.command,
-        params: cloneValue(cmd.params) as Record<string, unknown>,
-      }));
-
-      steps.push({
-        id: step.stepId,
-        description: `Recorded from run ${result.runId}`,
-        type: step.type,
-        commands,
-        ...(step.type === 'batch'
-          ? {
-              itemsPath: 'items',
-              continueOnItemFailure: true,
-            }
-          : {}),
-      } as WorkflowStepDefinition);
-    }
-
-    return {
-      version: config.version ?? 1,
-      name: config.name ?? `${result.workflowName}.recorded`,
-      description: config.description ?? `Recorded from workflow ${result.workflowName}`,
-      tags: config.tags,
-      steps,
-    };
-  }
-
-  private assertStarted(): void {
-    if (!this.draft) throw new Error('Recorder not started. Call start() first.');
-  }
-}
-
-// ============================================================================
 // Workflow Macro System
 // ============================================================================
 
@@ -1378,423 +1285,12 @@ export class WorkflowMacroSystem {
 }
 
 // ============================================================================
-// Prebuilt Workflow Definitions (DSL)
-// ============================================================================
-
-export const PREBUILT_WORKFLOWS: WorkflowDefinition[] = [
-  {
-    version: 1,
-    name: 'newco-sprint',
-    description: 'Research + validate + build MVP + test in one command.',
-    tags: ['newco', 'sprint', 'build'],
-    rollbackOnFailure: true,
-    steps: [
-      {
-        id: 'research',
-        description: 'Market and customer research.',
-        mode: 'parallel',
-        commands: [
-          { command: 'research.market-scan', retry: { maxAttempts: 2, backoffMs: 25 } },
-          { command: 'research.customer-interviews', retry: { maxAttempts: 2, backoffMs: 25 } },
-        ],
-      },
-      {
-        id: 'validate',
-        description: 'Validate problem/solution fit.',
-        commands: [{ command: 'validate.problem-solution-fit', retry: { maxAttempts: 2, backoffMs: 25 } }],
-      },
-      {
-        id: 'build-mvp',
-        description: 'Build MVP scaffold and prototype.',
-        commands: [
-          { command: 'build.mvp-scaffold' },
-          { command: 'build.prototype', retry: { maxAttempts: 2, backoffMs: 25 } },
-        ],
-      },
-      {
-        id: 'test-mvp',
-        description: 'Run lightweight user testing.',
-        commands: [{ command: 'test.user-testing' }],
-      },
-    ],
-  },
-  {
-    version: 1,
-    name: 'content-batch',
-    description: 'Write + edit + publish + promote a batch of content pieces.',
-    tags: ['content', 'batch'],
-    rollbackOnFailure: true,
-    defaults: {
-      contentPieces: [
-        { title: 'Piece 1', channel: 'blog' },
-        { title: 'Piece 2', channel: 'blog' },
-        { title: 'Piece 3', channel: 'newsletter' },
-        { title: 'Piece 4', channel: 'x' },
-        { title: 'Piece 5', channel: 'linkedin' },
-      ],
-    },
-    steps: [
-      {
-        id: 'content-plan',
-        description: 'Build topical and channel plan.',
-        commands: [{ command: 'content.plan-calendar' }],
-      },
-      {
-        id: 'produce-content-batch',
-        type: 'batch',
-        itemsPath: 'contentPieces',
-        concurrency: 2,
-        continueOnItemFailure: true,
-        commands: [
-          {
-            command: 'content.write',
-            params: {
-              title: '{{item.title}}',
-              channel: '{{item.channel}}',
-            },
-          },
-          {
-            command: 'content.edit',
-            params: {
-              title: '{{item.title}}',
-            },
-          },
-          {
-            command: 'content.publish',
-            params: {
-              title: '{{item.title}}',
-              channel: '{{item.channel}}',
-            },
-          },
-          {
-            command: 'content.promote',
-            params: {
-              title: '{{item.title}}',
-              channel: '{{item.channel}}',
-            },
-          },
-        ],
-      },
-    ],
-  },
-  {
-    version: 1,
-    name: 'deploy-check',
-    description: 'Build + test + security scan + deploy + verify.',
-    tags: ['deploy', 'ops', 'release'],
-    rollbackOnFailure: true,
-    steps: [
-      {
-        id: 'build',
-        commands: [{ command: 'deploy.build' }],
-      },
-      {
-        id: 'test',
-        commands: [{ command: 'deploy.test', retry: { maxAttempts: 2, backoffMs: 50 } }],
-      },
-      {
-        id: 'security-scan',
-        commands: [{ command: 'deploy.security-scan', retry: { maxAttempts: 2, backoffMs: 50 } }],
-      },
-      {
-        id: 'deploy',
-        onError: 'rollback',
-        commands: [
-          {
-            command: 'deploy.release',
-            rollback: {
-              command: 'deploy.rollback',
-            },
-          },
-        ],
-      },
-      {
-        id: 'verify',
-        commands: [{ command: 'deploy.verify' }],
-      },
-    ],
-  },
-  {
-    version: 1,
-    name: 'weekly-review',
-    description: 'Metrics + highlights + blockers + next-week plan.',
-    tags: ['review', 'planning'],
-    rollbackOnFailure: false,
-    steps: [
-      {
-        id: 'collect-metrics',
-        commands: [{ command: 'review.collect-metrics' }],
-      },
-      {
-        id: 'capture-highlights',
-        commands: [{ command: 'review.capture-highlights' }],
-      },
-      {
-        id: 'identify-blockers',
-        commands: [{ command: 'review.identify-blockers' }],
-      },
-      {
-        id: 'plan-next-week',
-        commands: [{ command: 'review.plan-next-week' }],
-      },
-    ],
-  },
-  {
-    version: 1,
-    name: 'incident-response',
-    description: 'Detect + triage + fix + postmortem with rollback safety.',
-    tags: ['incident', 'ops', 'response'],
-    rollbackOnFailure: true,
-    steps: [
-      {
-        id: 'detect',
-        commands: [{ command: 'incident.detect' }],
-      },
-      {
-        id: 'triage',
-        commands: [{ command: 'incident.triage' }],
-      },
-      {
-        id: 'fix',
-        onError: 'rollback',
-        commands: [
-          {
-            command: 'incident.fix',
-            retry: { maxAttempts: 3, backoffMs: 100 },
-            rollback: { command: 'incident.revert-fix' },
-          },
-        ],
-      },
-      {
-        id: 'postmortem',
-        commands: [{ command: 'incident.postmortem' }],
-      },
-    ],
-  },
-];
-
-// ============================================================================
 // Default Shortcut Command Library
 // ============================================================================
 
 export function createDefaultShortcutLibrary(config: { now?: () => Date } = {}): ShortcutCommandLibrary {
   const now = config.now ?? (() => new Date());
   const library = new ShortcutCommandLibrary();
-
-  const registerActivityCommand = (name: string, options: { idempotent?: boolean } = {}) => {
-    library.registerCommand({
-      name,
-      description: `${name} command`,
-      idempotent: options.idempotent ?? true,
-      async handler(params, ctx) {
-        if (params.forceFail === true || params.fail === true) {
-          throw new Error(`${name} forced failure`);
-        }
-
-        const timeline = Array.isArray(ctx.state.timeline)
-          ? (ctx.state.timeline as Array<Record<string, unknown>>)
-          : [];
-
-        const event = {
-          command: name,
-          stepId: ctx.stepId,
-          attempt: ctx.attempt,
-          item: ctx.item,
-          params,
-          timestamp: now().toISOString(),
-        };
-
-        timeline.push(event);
-
-        return {
-          output: event,
-          stateUpdates: {
-            timeline,
-            lastCommand: name,
-            lastStepId: ctx.stepId,
-          },
-        };
-      },
-      async rollback(params, ctx) {
-        const timeline = Array.isArray(ctx.state.timeline)
-          ? (ctx.state.timeline as Array<Record<string, unknown>>)
-          : [];
-
-        timeline.push({
-          command: `${name}:rollback`,
-          stepId: ctx.stepId,
-          params,
-          timestamp: now().toISOString(),
-        });
-
-        ctx.state.timeline = timeline;
-      },
-    });
-  };
-
-  [
-    // NewCo sprint commands
-    'research.market-scan',
-    'research.customer-interviews',
-    'validate.problem-solution-fit',
-    'build.mvp-scaffold',
-    'build.prototype',
-    'test.user-testing',
-
-    // Content batch commands
-    'content.plan-calendar',
-    'content.write',
-    'content.edit',
-    'content.publish',
-    'content.promote',
-
-    // Deploy commands
-    'deploy.build',
-    'deploy.test',
-    'deploy.security-scan',
-    'deploy.release',
-    'deploy.verify',
-    'deploy.rollback',
-
-    // Weekly review commands
-    'review.collect-metrics',
-    'review.capture-highlights',
-    'review.identify-blockers',
-    'review.plan-next-week',
-
-    // Incident commands
-    'incident.detect',
-    'incident.triage',
-    'incident.fix',
-    'incident.revert-fix',
-    'incident.postmortem',
-  ].forEach((commandName) => registerActivityCommand(commandName));
-
-  library.registerShortcut('newco.research', 'research.market-scan');
-  library.registerShortcut('newco.interview', 'research.customer-interviews');
-  library.registerShortcut('content.create', 'content.write');
-  library.registerShortcut('content.go-live', 'content.publish');
-  library.registerShortcut('deploy.go', 'deploy.release');
-  library.registerShortcut('incident.mitigate', 'incident.fix');
-
+  registerDefaultCommandsAndShortcuts(library, now);
   return library;
-}
-
-// ============================================================================
-// Helpers
-// ============================================================================
-
-function estimateTotalUnits(workflow: WorkflowDefinition, inputs: Record<string, unknown>): number {
-  let units = 0;
-
-  for (const step of workflow.steps) {
-    if (step.type === 'batch') {
-      const list = resolveBatchItems(step, inputs);
-      units += Math.max(1, list.length);
-      continue;
-    }
-
-    units += 1;
-  }
-
-  return Math.max(1, units);
-}
-
-function resolveBatchItems(step: WorkflowBatchStepDefinition, inputs: Record<string, unknown>): unknown[] {
-  const raw = getPathValue(inputs, step.itemsPath);
-  if (!Array.isArray(raw)) return [];
-  return raw;
-}
-
-function evaluateCondition(
-  condition: WorkflowStepDefinition['condition'],
-  context: WorkflowEvaluationContext
-): boolean {
-  if (condition === undefined) return true;
-
-  if (typeof condition === 'function') {
-    return Boolean(condition(context));
-  }
-
-  const raw = resolveTemplateValue(condition, {
-    inputs: context.inputs,
-    state: context.state,
-    workflow: context.workflowName,
-  });
-
-  if (typeof raw === 'boolean') return raw;
-  if (typeof raw === 'number') return raw !== 0;
-  if (typeof raw === 'string') {
-    const normalized = raw.trim().toLowerCase();
-    if (normalized === 'false' || normalized === '0' || normalized === 'no' || normalized === '') return false;
-    return true;
-  }
-
-  return Boolean(raw);
-}
-
-function resolveTemplateValue(value: unknown, context: Record<string, unknown>): unknown {
-  if (typeof value === 'string') {
-    const whole = value.match(/^\{\{\s*([^{}]+?)\s*\}\}$/);
-    if (whole) {
-      const lookup = getPathValue(context, whole[1]);
-      return lookup;
-    }
-
-    return value.replace(/\{\{\s*([^{}]+?)\s*\}\}/g, (_, path: string) => {
-      const lookup = getPathValue(context, path);
-      if (lookup === null || lookup === undefined) return '';
-      return String(lookup);
-    });
-  }
-
-  if (Array.isArray(value)) {
-    return value.map((item) => resolveTemplateValue(item, context));
-  }
-
-  if (isPlainObject(value)) {
-    const out: Record<string, unknown> = {};
-    for (const [k, v] of Object.entries(value)) {
-      out[k] = resolveTemplateValue(v, context);
-    }
-    return out;
-  }
-
-  return value;
-}
-
-function getPathValue(source: unknown, path: string): unknown {
-  if (!path) return undefined;
-
-  const parts = path.split('.').map((part) => part.trim()).filter(Boolean);
-  let current: unknown = source;
-
-  for (const part of parts) {
-    if (!isObjectLike(current)) return undefined;
-
-    if (Array.isArray(current)) {
-      const idx = Number(part);
-      if (!Number.isInteger(idx)) return undefined;
-      current = current[idx];
-      continue;
-    }
-
-    current = (current as Record<string, unknown>)[part];
-  }
-
-  return current;
-}
-
-function isObjectLike(value: unknown): value is Record<string, unknown> {
-  return typeof value === 'object' && value !== null;
-}
-
-function isPlainObject(value: unknown): value is Record<string, unknown> {
-  if (!isObjectLike(value)) return false;
-  return Object.getPrototypeOf(value) === Object.prototype || Object.getPrototypeOf(value) === null;
-}
-
-function cloneValue<T>(value: T): T {
-  if (value === undefined) return value;
-  return JSON.parse(JSON.stringify(value)) as T;
 }
